Compronents structure reorganized
