--- frontend/src/components/admin/ManageData.tsx
@@ -1,37 +1,0 @@
-import React from 'react';
-
-export const ManageData: React.FC = () => {
-  return (
-    <div className="p-6">
-      <div className="flex justify-between items-center mb-6">
-        <h2 className="text-2xl font-bold text-gray-900">Data Management</h2>
-        <button className="bg-blue-600 text-white px-4 py-2 rounded-md hover:bg-blue-700 transition duration-200">
-          Export Data
-        </button>
-      </div>
-
-      {/* Placeholder content - will be implemented when backend data endpoints are ready */}
-      <div className="bg-gray-50 rounded-lg p-8 text-center">
-        <div className="text-6xl mb-4">📊</div>
-        <h3 className="text-xl font-semibold text-gray-700 mb-2">Data Management</h3>
-        <p className="text-gray-600 mb-4">
-          This section will provide analytics, reports, and data management tools for your restaurant.
-        </p>
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-4 mt-8">
-          <div className="bg-white p-4 rounded-lg shadow">
-            <h4 className="font-semibold text-gray-800 mb-2">Analytics</h4>
-            <p className="text-sm text-gray-600">View sales and performance metrics</p>
-          </div>
-          <div className="bg-white p-4 rounded-lg shadow">
-            <h4 className="font-semibold text-gray-800 mb-2">Reports</h4>
-            <p className="text-sm text-gray-600">Generate detailed business reports</p>
-          </div>
-          <div className="bg-white p-4 rounded-lg shadow">
-            <h4 className="font-semibold text-gray-800 mb-2">Data Export</h4>
-            <p className="text-sm text-gray-600">Export data for external analysis</p>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
